함수명 재지정, fifteenPuzzle 클래스 파일 분리
