+++ src/debugging.js
@@ -1,7 +1,7 @@
 const prettierShowArr = (twoDimensionArr, string) => {
   if(string != null) console.log(string);
   for(let i = 0; i < 4; i++) {
-    console.log(twoDimensionArr[i]);
+    console.log(...twoDimensionArr[i]);
   }
 }
 export default prettierShowArr;
